игра в кнб начало.

+++ Version_1/handlerss/client.py
@@ -6,12 +6,14 @@ from Version_1.keyboards.keyboard import kb_menu, kb_some
 
 users = {}
 players = {}
+play = False
 
 
 def who_win(players: dict):
     pass
 
 
+@dp.message_handler(commands=['start'])
 async def start(message: Message):
     global users
     text = f'Добро пожаловать {message.from_user.full_name}'
@@ -19,33 +21,70 @@ async def start(message: Message):
         'id': message.from_user.id,
         'username': message.from_user.username
     }
-    print(users)
-    print(message.from_user)
-    print(message.chat)
+    # print(users)
+    # print(message.from_user)
+    # print(message.chat)
+    print(players)
     await bot.send_message(chat_id=message.chat.id, text=text)
 
 
+@dp.message_handler(commands=['kmn'])
 async def game(message: Message):
+    global play
+    play = True
     await message.answer(text='Выберите игроков!!')
-    await message.answer(text='Старт игры.\n'
-                              'Напишите в лс боту вам выбор:\n'
+
+
+@dp.message_handler(commands=['play'])
+async def choice(message: Message):
+    global play
+    if not play:
+        await message.answer(text='игра еще не началась')
+    await message.answer(text='игра началась')
+    await message.answer(text='выберите что ответить:\n'
                               'к - камень\n'
                               'н - ножницы\n'
                               'б - бумага')
+    text = message.text
+    if text not in ['к', 'н', 'б']:
+        await message.answer(text='такого ответ нет. \n'
+                                  'нажмите /play и выберите еще раз')
+    else:
+        global players
+        players[message.from_user.id].append(text)
 
 
+
+@dp.message_handler(commands=['win'])
+async def who_win(message: Message):
+    pass
+
+
+@dp.message_handler(commands=['menu'])
 async def menu(message: Message):
     await message.answer(text='выбор', reply_markup=kb_menu)
 
 
+@dp.message_handler(commands=['some'])
 async def somecommand(message: Message):
     await message.answer(text='выбор', reply_markup=kb_some)
 
 
+@dp.message_handler()
 async def commands(message: Message):
-    text_message = message.text
-    await bot.send_message(chat_id=message.chat.id, text=text_message)
-    await message.delete()
+    if play:
+        pl = message.text.split()
+        global players
+        u = [value['username'] for key, value in users.items()]
+        for p in pl:
+            if p in u:
+                players[message.from_user.id] = [p]
+            else:
+                await message.answer(text=f'такого пользователя нет {p}')
+    else:
+        text_message = message.text
+        await bot.send_message(chat_id=message.chat.id, text=text_message)
+        await message.delete()
 
 #
 # #     if {i.lower().translate(str.maketrans('', '', string.punctuation)) for i in message.text.split(' ')} \
